backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.20
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic[email]==2.4.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
passlib==1.7.4
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.7.2
beautifulsoup4==4.12.3
requests==2.31.0
scikit-learn==1.5.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pandas==2.2.2
matplotlib==3.9.2
pytest==8.3.3
